pyproject.toml
[tool.poetry]
name = "django-generic-tasks"
version = "0.3.0"
description = ""
authors = ["Jiayu Yi <yijiayu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django-ninja = "^0.17.0"
Django = "^4.0.4"
google-auth = "^2.6.6"
google-cloud-tasks = "^2.8.1"
CacheControl = "^0.12.11"
requests = "^2.27.1"
pydantic = "^1.9.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"
isort = "^5.10.1"
pre-commit = "^2.18.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
